feat: concepts refactor
