+++ src/concepts/mod.rs
@@ -1,0 +1,6 @@
+//! Key concepts for Instrumentality.
+
+pub mod data;
+pub mod group;
+pub mod subject;
+pub mod user;
