+++ src/lib.rs
@@ -15,13 +15,10 @@
 //! [Axum]: https://github.com/tokio-rs/axum/
 
 pub mod boot;
+pub mod concepts;
 pub mod config;
-pub mod data;
 pub mod database;
-pub mod group;
 #[macro_use]
 pub mod routes;
 pub mod server;
-pub mod subject;
-pub mod user;
 pub mod utils;
