+++ src/routes/groups/delete.rs
@@ -10,11 +10,11 @@ use mongodb::bson::doc;
 use mongodb::Collection;
 use serde::{Deserialize, Serialize};
 
+use crate::concepts::group::Group;
+use crate::concepts::user::User;
 use crate::database::DBHandle;
-use crate::group::Group;
 use crate::routes::response::ErrorResponse;
 use crate::routes::response::OkResponse;
-use crate::user::User;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct DeleteGroupRequest {
