+++ src/routes/groups/update.rs
@@ -10,11 +10,11 @@ use mongodb::bson::doc;
 use mongodb::{bson, Collection};
 use serde::{Deserialize, Serialize};
 
+use crate::concepts::group::Group;
+use crate::concepts::subject::*;
+use crate::concepts::user::User;
 use crate::database::DBHandle;
-use crate::group::Group;
 use crate::routes::response::{ErrorResponse, OkResponse};
-use crate::subject::*;
-use crate::user::User;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct UpdateGroupRequest {
