+++ src/routes/subjects/delete.rs
@@ -10,13 +10,13 @@ use mongodb::bson::doc;
 use mongodb::Collection;
 use serde::{Deserialize, Serialize};
 
+use crate::concepts::group::Group;
+use crate::concepts::subject::*;
+use crate::concepts::user::User;
 use crate::database::DBHandle;
-use crate::group::Group;
 use crate::routes::queue;
 use crate::routes::response::ErrorResponse;
 use crate::routes::response::OkResponse;
-use crate::subject::*;
-use crate::user::User;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct DeleteSubjectRequest {
