+++ src/routes/user/mod.rs
@@ -1,4 +1,5 @@
 //! Routes for user.
 
+pub mod from_request_parts;
 pub mod login;
 pub mod reset;
