+++ src/utils/mod.rs
@@ -1,2 +1,3 @@
+//! Common utilities for Instrumentality.
 pub mod deserialise_array;
 pub mod random;
